modularize artwork group
